requirements.txt
Pillow==2.2.1
argparse==1.2.1
distribute==0.6.24
selenium==2.37.2
wsgiref==0.1.2
